chore: rename benchmark files to better names
