+++ test/benchmarks/002-transaction-and-span-overhead-realistic-size.js
@@ -2,23 +2,32 @@
 
 process.on('SIGUSR2', end)
 
+const benchName = 'transaction-and-span-overhead-realistic-size'
+
 let agent
 if (process.env.AGENT) {
   agent = require('../../').start({
-    serviceName: '001-custom-transactions',
-    captureExceptions: false,
-    captureSpanStackTraces: false
+    serviceName: benchName,
+    captureExceptions: false
   })
 }
+
+const callstack = require('./utils/callstack')
 
 let start
 let stop = false
 const pid = process.argv[2]
 const warmup = 1e4
-const runtime = 10
+const runtime = 20
 const metrics = {
   transactions: 0
 }
+
+// To avoid randomness, but still generate what appears to be natural random
+// call stacks, number of spans etc, use a pre-defined set of numbers
+const numbers = [2, 5, 10, 1, 2, 21, 2, 5, 6, 9, 1, 11, 9, 8, 12]
+let numbersSpanIndex = 5
+let numbersStackLevelIndex = 0
 
 // warmup
 console.error('Warming up for %d transactions...', warmup)
@@ -42,10 +51,14 @@ addTransaction('warmup', function runAgain () {
 function addTransaction (name, cb) {
   if (stop) return
   if (agent) agent.startTransaction(name)
-  addSpan(name, () => {
-    if (agent) agent.endTransaction()
-    metrics.transactions++
-    setImmediate(cb)
+  const amount = numbers[numbersStackLevelIndex++ % numbers.length]
+  callstack(amount, () => {
+    const amount = numbers[numbersSpanIndex++ % numbers.length]
+    addSpan(name, amount, () => {
+      if (agent) agent.endTransaction()
+      metrics.transactions++
+      setImmediate(cb)
+    })
   })
 }
 
@@ -77,7 +90,7 @@ function end () {
 function shutdown (duration) {
   if (pid) process.kill(pid, 'SIGUSR2')
   process.stdout.write(JSON.stringify({
-    name: 'custom-transactions',
+    name: benchName,
     warmup: { count: warmup, unit: 'transactions' },
     duration,
     metrics
