+++ test/benchmarks/003-transaction-and-span-with-stack-trace.js
@@ -2,30 +2,34 @@
 
 process.on('SIGUSR2', end)
 
+const benchName = 'transaction-and-span-with-stack-trace'
+
 let agent
 if (process.env.AGENT) {
   agent = require('../../').start({
-    serviceName: '002-custom-transactions-normal-sized',
+    serviceName: benchName,
     captureExceptions: false
   })
+  // v1 no-ops
+  if (agent._instrumentation._queue) {
+    agent._instrumentation._queue.add = noop
+  }
+  // v2 no-ops
+  if (agent._apmServer) {
+    agent._apmServer.sendTransaction = noop
+    agent._apmServer.sendSpan = noop
+    agent._apmServer.sendError = noop
+  }
 }
-
-const callstack = require('./utils/callstack')
 
 let start
 let stop = false
 const pid = process.argv[2]
 const warmup = 1e4
-const runtime = 20
+const runtime = 10
 const metrics = {
   transactions: 0
 }
-
-// To avoid randomness, but still generate what appears to be natural random
-// call stacks, number of spans etc, use a pre-defined set of numbers
-const numbers = [2, 5, 10, 1, 2, 21, 2, 5, 6, 9, 1, 11, 9, 8, 12]
-let numbersSpanIndex = 5
-let numbersStackLevelIndex = 0
 
 // warmup
 console.error('Warming up for %d transactions...', warmup)
@@ -49,14 +53,10 @@ addTransaction('warmup', function runAgain () {
 function addTransaction (name, cb) {
   if (stop) return
   if (agent) agent.startTransaction(name)
-  const amount = numbers[numbersStackLevelIndex++ % numbers.length]
-  callstack(amount, () => {
-    const amount = numbers[numbersSpanIndex++ % numbers.length]
-    addSpan(name, amount, () => {
-      if (agent) agent.endTransaction()
-      metrics.transactions++
-      setImmediate(cb)
-    })
+  addSpan(name, () => {
+    if (agent) agent.endTransaction()
+    metrics.transactions++
+    setImmediate(cb)
   })
 }
 
@@ -88,10 +88,12 @@ function end () {
 function shutdown (duration) {
   if (pid) process.kill(pid, 'SIGUSR2')
   process.stdout.write(JSON.stringify({
-    name: 'custom-transactions-normal-sized',
+    name: benchName,
     warmup: { count: warmup, unit: 'transactions' },
     duration,
     metrics
   }))
   process.exit()
 }
+
+function noop () {}
